requirements.txt
aiogram==3.7.0
telethon==1.36.0
APScheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
